# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApproveRequest": {
        "properties": {
          "attestation": {
            "minLength": 1,
            "title": "Attestation",
            "type": "string"
          },
          "expected_revision": {
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision",
          "attestation"
        ],
        "title": "ApproveRequest",
        "type": "object"
      },
      "BlockedInfo": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "rule_name": {
            "title": "Rule Name",
            "type": "string"
          },
          "rule_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Version Id"
          }
        },
        "required": [
          "rule_version_id",
          "rule_name",
          "message"
        ],
        "title": "BlockedInfo",
        "type": "object"
      },
...
      "CaseSubmission": {
        "description": "POST /plan-options body.",
        "properties": {
          "assessment_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Time"
          },
          "case_ref": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Case Ref",
            "type": "string"
          },
          "catalog_release_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Release Id"
          },
          "intake": {
            "$ref": "#/components/schemas/Intake"
          },
          "narrative": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative"
          },
          "service_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Date"
          }
        },
        "required": [
          "case_ref"
        ],
        "title": "CaseSubmission",
        "type": "object"
      },
...
      "DraftPatch": {
        "properties": {
          "edit_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edit Reason"
          },
          "expected_revision": {
            "title": "Expected Revision",
            "type": "integer"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PlanItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "selected_option_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Option Id"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "DraftPatch",
        "type": "object"
      },
      "ExcludedCandidate": {
        "properties": {
          "protocol_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol Version Id"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "rule_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Version Id"
          },
          "variant_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Version Id"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ExcludedCandidate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Intake": {
        "properties": {
          "fields": {
            "additionalProperties": {
              "$ref": "#/components/schemas/IntakeField"
            },
            "title": "Fields",
            "type": "object"
          }
        },
        "title": "Intake",
        "type": "object"
      },
...
      "MissingField": {
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "required_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required By"
          }
        },
        "required": [
          "field",
          "reason"
        ],
        "title": "MissingField",
        "type": "object"
      },
      "PathwayPreview": {
        "description": "Conditional, time-based preview shown before individual prescribing is possible (spec \u00a720B).",
        "properties": {
          "approval_state": {
            "title": "Approval State",
            "type": "string"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "branch_conditions": {
            "items": {
              "type": "string"
            },
            "title": "Branch Conditions",
            "type": "array"
          },
          "condition": {
            "title": "Condition",
            "type": "string"
          },
          "label": {
            "default": "Protocol preview\u2014requires assessment and PT approval",
            "title": "Label",
            "type": "string"
          },
          "protocol_name": {
            "title": "Protocol Name",
            "type": "string"
          },
          "protocol_version_id": {
            "title": "Protocol Version Id",
            "type": "string"
          },
          "recovery_horizon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery Horizon"
          },
          "source_references": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "title": "Source References",
            "type": "array"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineRow"
            },
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "protocol_version_id",
          "protocol_name",
          "condition",
          "approval_state"
        ],
        "title": "PathwayPreview",
        "type": "object"
      },
...
      "PlanItem": {
        "properties": {
          "cautions": {
            "items": {
              "type": "string"
            },
            "title": "Cautions",
            "type": "array"
          },
          "clinical_use_version_id": {
            "title": "Clinical Use Version Id",
            "type": "string"
          },
          "dose_label": {
            "default": "protocol_example",
            "title": "Dose Label",
            "type": "string"
          },
          "evidence_claim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Claim Ids",
            "type": "array"
          },
          "goal_ids": {
            "items": {
              "type": "string"
            },
            "title": "Goal Ids",
            "type": "array"
          },
          "instructions": {
            "items": {
              "type": "string"
            },
            "title": "Instructions",
            "type": "array"
          },
          "media_state": {
            "default": "not_requested",
            "title": "Media State",
            "type": "string"
          },
          "media_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Version Id"
          },
          "prescribed_dose": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Dose"
              },
              {
                "type": "null"
              }
            ]
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "source_references": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "title": "Source References",
            "type": "array"
          },
          "variant_name": {
            "title": "Variant Name",
            "type": "string"
          },
          "variant_version_id": {
            "title": "Variant Version Id",
            "type": "string"
          }
        },
        "required": [
          "variant_version_id",
          "variant_name",
          "clinical_use_version_id"
        ],
        "title": "PlanItem",
        "type": "object"
      },
      "PlanOption": {
        "properties": {
          "approval": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval"
          },
          "case_snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Snapshot Id"
          },
          "catalog_release_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Release Id"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "differences_from_source": {
            "items": {
              "type": "string"
            },
            "title": "Differences From Source",
            "type": "array"
          },
          "equipment": {
            "items": {
              "type": "string"
            },
            "title": "Equipment",
            "type": "array"
          },
          "estimated_session_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Session Minutes"
          },
          "evidence_limitations": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Limitations",
            "type": "array"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlanItem"
            },
            "title": "Items",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "monitoring_rules": {
            "items": {
              "type": "string"
            },
            "title": "Monitoring Rules",
            "type": "array"
          },
          "option_id": {
            "title": "Option Id",
            "type": "string"
          },
          "progression_requirements": {
            "items": {
              "type": "string"
            },
            "title": "Progression Requirements",
            "type": "array"
          },
          "protocol_version_ids": {
            "items": {
              "type": "string"
            },
            "title": "Protocol Version Ids",
            "type": "array"
          },
          "rank_explanation": {
            "additionalProperties": true,
            "title": "Rank Explanation",
            "type": "object"
          },
          "reassessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Reassessment"
              },
              {
                "type": "null"
              }
            ]
          },
          "rule_version_ids": {
            "items": {
              "type": "string"
            },
            "title": "Rule Version Ids",
            "type": "array"
          },
          "status": {
            "default": "draft",
            "title": "Status",
            "type": "string"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineRow"
            },
            "title": "Timeline",
            "type": "array"
          },
          "uncovered_goals": {
            "items": {
              "type": "string"
            },
            "title": "Uncovered Goals",
            "type": "array"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationResult"
          }
        },
        "required": [
          "option_id",
          "label"
        ],
        "title": "PlanOption",
        "type": "object"
      },
      "PlanOptionsResponse": {
        "properties": {
          "assignable": {
            "default": false,
            "title": "Assignable",
            "type": "boolean"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "blocked": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlockedInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "candidate_conditions": {
            "items": {
              "type": "string"
            },
            "title": "Candidate Conditions",
            "type": "array"
          },
          "case_revision": {
            "default": 1,
            "title": "Case Revision",
            "type": "integer"
          },
          "case_snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Snapshot Id"
          },
          "catalog_release_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Release Id"
          },
          "excluded_candidates": {
            "items": {
              "$ref": "#/components/schemas/ExcludedCandidate"
            },
            "title": "Excluded Candidates",
            "type": "array"
          },
          "missing_fields": {
            "items": {
              "$ref": "#/components/schemas/MissingField"
            },
            "title": "Missing Fields",
            "type": "array"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/PlanOption"
            },
            "title": "Options",
            "type": "array"
          },
          "pathway_previews": {
            "items": {
              "$ref": "#/components/schemas/PathwayPreview"
            },
            "title": "Pathway Previews",
            "type": "array"
          },
          "phase_windows": {
            "items": {
              "$ref": "#/components/schemas/TimeWindow"
            },
            "title": "Phase Windows",
            "type": "array"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "prescription": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Prescription"
          },
          "requires_pt_review": {
            "default": true,
            "title": "Requires Pt Review",
            "type": "boolean"
          },
          "segment_explanations": {
            "items": {
              "$ref": "#/components/schemas/SegmentExplanation"
            },
            "title": "Segment Explanations",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/PlanStatus"
          },
          "timeline_anchor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeline Anchor"
          }
        },
        "required": [
          "status"
        ],
        "title": "PlanOptionsResponse",
        "type": "object"
      },
      "PlanStatus": {
        "enum": [
          "needs_assessment",
          "blocked_for_clinical_review",
          "draft_ready"
        ],
        "title": "PlanStatus",
        "type": "string"
      },
...
      "Reassessment": {
        "properties": {
          "checks": {
            "items": {
              "type": "string"
            },
            "title": "Checks",
            "type": "array"
          },
          "provenance": {
            "default": "unknown",
            "title": "Provenance",
            "type": "string"
          },
          "timing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timing"
          }
        },
        "title": "Reassessment",
        "type": "object"
      },
      "ReassessmentRequest": {
        "properties": {
          "assessment_time": {
            "title": "Assessment Time"
          },
          "intake": {
            "title": "Intake"
          },
          "narrative": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative"
          }
        },
        "required": [
          "intake"
        ],
        "title": "ReassessmentRequest",
        "type": "object"
      },
...
      "SegmentExplanation": {
        "properties": {
          "classification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification"
          },
          "classification_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification Version"
          },
          "dimension": {
            "title": "Dimension",
            "type": "string"
          },
          "effect": {
            "default": "recorded for context; no supported modification applied",
            "title": "Effect",
            "type": "string"
          },
          "evidence_gap": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Gap"
          },
          "observed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed"
          },
          "rule_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Version Id"
          }
        },
        "required": [
          "dimension",
          "observed"
        ],
        "title": "SegmentExplanation",
        "type": "object"
      },
...
      "TimeWindow": {
        "properties": {
          "anchor": {
            "title": "Anchor",
            "type": "string"
          },
          "end": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "provenance": {
            "default": "unknown",
            "title": "Provenance",
            "type": "string"
          },
          "provisional": {
            "default": true,
            "title": "Provisional",
            "type": "boolean"
          },
          "start": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "unit": {
            "default": "weeks",
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "anchor"
        ],
        "title": "TimeWindow",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MoveAI Exercise Ingestion and Plan-Drafting Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/plan-options": {
      "post": {
        "operationId": "create_plan_options_v1_plan_options_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOptionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Plan Options",
        "tags": [
          "plans"
        ]
      }
    },
    "/v1/plans/{plan_id}": {
      "get": {
        "operationId": "get_plan_v1_plans__plan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Plan V1 Plans  Plan Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Plan",
        "tags": [
          "plans"
        ]
      }
    },
    "/v1/plans/{plan_id}/approve": {
      "post": {
        "operationId": "approve_plan_v1_plans__plan_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Plan V1 Plans  Plan Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Plan",
        "tags": [
          "plans"
        ]
      }
    },
    "/v1/plans/{plan_id}/approved": {
      "get": {
        "operationId": "approved_plan_v1_plans__plan_id__approved_get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approved Plan V1 Plans  Plan Id  Approved Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approved Plan",
        "tags": [
          "plans"
        ]
      }
    },
    "/v1/plans/{plan_id}/draft": {
      "patch": {
        "operationId": "edit_draft_v1_plans__plan_id__draft_patch",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Edit Draft V1 Plans  Plan Id  Draft Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Draft",
        "tags": [
          "plans"
        ]
      }
    },
    "/v1/plans/{plan_id}/reassessments": {
      "post": {
        "operationId": "create_reassessment_v1_plans__plan_id__reassessments_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOptionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Reassessment",
        "tags": [
          "plans"
        ]
      }
    },
...
  }
}